refactor(payment): refactor zustand into checkoutstore
