+++ src/containers/AccountModal/forms/Checkout.tsx
@@ -2,16 +2,17 @@ import React, { useCallback, useEffect, useState } from 'react';
 import { useHistory } from 'react-router';
 import { useTranslation } from 'react-i18next';
 
-import CheckoutForm from '../../../components/CheckoutForm/CheckoutForm';
-import { CheckoutStore, createOrder, updateOrder, getPaymentMethods, paymentWithoutDetails, adyenPayment, paypalPayment } from '../../../stores/CheckoutStore';
-import { addQueryParam } from '../../../utils/history';
-import useForm from '../../../hooks/useForm';
-import LoadingOverlay from '../../../components/LoadingOverlay/LoadingOverlay';
-import Adyen from '../../../components/Adyen/Adyen';
-import PayPal from '../../../components/PayPal/PayPal';
-import NoPaymentRequired from '../../../components/NoPaymentRequired/NoPaymentRequired';
-import { addQueryParams } from '../../../utils/formatting';
-import { ConfigStore } from '../../../stores/ConfigStore';
+import { addQueryParam } from '#src/utils/history';
+import useForm from '#src/hooks/useForm';
+import { addQueryParams } from '#src/utils/formatting';
+import { ConfigStore } from '#src/stores/ConfigStore';
+import CheckoutForm from '#src/components/CheckoutForm/CheckoutForm';
+import LoadingOverlay from '#src/components/LoadingOverlay/LoadingOverlay';
+import Adyen from '#src/components/Adyen/Adyen';
+import PayPal from '#src/components/PayPal/PayPal';
+import NoPaymentRequired from '#src/components/NoPaymentRequired/NoPaymentRequired';
+import { useCheckoutStore } from '#src/stores/CheckoutStore';
+import { createOrder, updateOrder, getPaymentMethods, paymentWithoutDetails, adyenPayment, paypalPayment } from '#src/stores/CheckoutController';
 import { reloadActiveSubscription } from '#src/stores/AccountController';
 
 const Checkout = () => {
@@ -24,7 +25,7 @@ const Checkout = () => {
   const [couponCodeApplied, setCouponCodeApplied] = useState(false);
   const [paymentMethodId, setPaymentMethodId] = useState<number | undefined>(undefined);
 
-  const { order, offer, paymentMethods } = CheckoutStore.useState((s) => s);
+  const [order, offer, paymentMethods] = useCheckoutStore((s) => [s.order, s.offer, s.paymentMethods]);
 
   const couponCodeForm = useForm({ couponCode: '' }, async (values, { setSubmitting, setErrors }) => {
     setUpdatingOrder(true);
